week00/code/embedding/requirement.txt
Package            Version
------------------ ---------
annotated-types    0.7.0
anyio              4.9.0
cachetools         5.5.2
certifi            2025.4.26
charset-normalizer 3.4.2
google-auth        2.39.0
google-genai       1.13.0
h11                0.16.0
httpcore           1.0.9
httpx              0.28.1
idna               3.10
pip                24.0
pyasn1             0.6.1
pyasn1_modules     0.4.2
pydantic           2.11.4
pydantic_core      2.33.2
python-dotenv      1.1.0
requests           2.32.3
rsa                4.9.1
sniffio            1.3.1
typing_extensions  4.13.2
typing-inspection  0.4.0
urllib3            2.4.0
websockets         15.0.1
